# apps/cli/templates/monorepo/apps/api/static/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
...
    "/api/v1/auth/google/callback": {
      "get": {
        "description": "Handle Google OAuth callback and redirect",
        "summary": "Google login callback",
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "auth.googleCallback",
        "responses": {
          "302": {
            "description": "302",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "description": "Verify user email using a verification code",
        "summary": "Verify email",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "operationId": "auth.verifyEmail",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "code": {
                    "type": "string",
                    "minLength": 4,
                    "maxLength": 10
                  }
                },
                "required": [
                  "email",
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "username": {
                      "type": "string",
                      "minLength": 3,
                      "maxLength": 50
                    },
                    "googleId": {
                      "type": "string"
                    },
                    "emailVerifiedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "lastLoginAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "isAdmin": {
                      "default": false,
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "deletedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "email",
                    "username",
                    "id",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        401
                      ]
                    },
                    "message": {
                      "default": "Sorry, you are not authorized to access this resource.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "403",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        403
                      ]
                    },
                    "message": {
                      "default": "Sorry, you do not have permission to access this resource.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        404
                      ]
                    },
                    "message": {
                      "default": "The requested resource was not found.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "500",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        500
                      ]
                    },
                    "message": {
                      "default": "Sorry, something went wrong on our end. Please try again later.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Refresh access using the refresh cookie",
        "summary": "Refresh session",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "operationId": "auth.refresh",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "username": {
                      "type": "string",
                      "minLength": 3,
                      "maxLength": 50
                    },
                    "googleId": {
                      "type": "string"
                    },
                    "emailVerifiedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "lastLoginAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "isAdmin": {
                      "default": false,
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "deletedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "email",
                    "username",
                    "id",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        401
                      ]
                    },
                    "message": {
                      "default": "Sorry, you are not authorized to access this resource.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "403",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        403
                      ]
                    },
                    "message": {
                      "default": "Sorry, you do not have permission to access this resource.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        404
                      ]
                    },
                    "message": {
                      "default": "The requested resource was not found.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "500",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        500
                      ]
                    },
                    "message": {
                      "default": "Sorry, something went wrong on our end. Please try again later.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/resend-verification": {
      "post": {
        "description": "Resend the email verification code",
        "summary": "Resend verification",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "operationId": "auth.resendVerification",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "default": 200,
                      "type": "integer"
                    },
                    "message": {
                      "default": "Request processed successfully.",
                      "type": "string"
                    },
                    "success": {
                      "default": true,
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "401",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        401
                      ]
                    },
                    "message": {
                      "default": "Sorry, you are not authorized to access this resource.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "403",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        403
                      ]
                    },
                    "message": {
                      "default": "Sorry, you do not have permission to access this resource.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        404
                      ]
                    },
                    "message": {
                      "default": "The requested resource was not found.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "500",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        500
                      ]
                    },
                    "message": {
                      "default": "Sorry, something went wrong on our end. Please try again later.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "version": "1.0.0",
    "title": "go-kickstart REST API - Documentation",
    "description": "go-kickstart REST API - Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "name": "access_token",
        "in": "cookie"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}